week 4 end video, router refactored
