+++ 02-express-tutorial/routes/auth.js
@@ -4,7 +4,7 @@ const router = express.Router();
 router.post("/", (req, res) => {
 	const { name } = req.body;
 	if (name) {
-		return res.status(201).send(`Welcome ${name}`);
+		return res.status(200).send(`Welcome ${name}`);
 	}
 	res.status(401).send(`please provide credentials`);
 });
